moved module address query
